OMImage: Fixed a inconsistency where it was used short instead of int in constructor OMList: Fixed a bug with subitems not updating text texture HelperFunctions: Added methods to get misc plugin info (plugins, types, names, icons etc...) IHighLevel: Added attribute for PluginLevel (PluginType definition for HighLevel plugins) used to filter plugins. Graphics: Fixed a small bug in GenerateTextTexture that caused it to fail if we wanted to write to a new empty bitmap. Graphics: Modified smoothingMode for textdrawing to remove the "jagged" edges when drawing fonts. OImage: Added option to generate image from font Misc plugins: Added PluginIcon and PluginLevel attribute Added support for generating pluginIcon based on WebDings font (if first character is * then it will use the rest of the string as an icon) MainMenu: Added support for pluginIcons MainMenu: Updated the "look and feel" of the button assignment menu (also added in icon support here) OMDialog: Added support for configurable button setup. OMDebug: Modefied time initialization so that time vales will be logged from startup of plugin.

+++ FrontEnd/OpenMobile.Framework/Controls/OMImage.cs
@@ -69,7 +69,7 @@ namespace OpenMobile.Controls
         /// <param name="top"></param>
         /// <param name="width"></param>
         /// <param name="height"></param>
-        public OMImage(short left, short top, short width, short height)
+        public OMImage(int left, int top, int width, int height)
         {
             this.top = top;
             this.left = left;
